requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pydantic>=2.4.0
sqlmodel>=0.0.11
sqlalchemy>=2.0.0
pillow>=10.0.0
python-dotenv>=1.0.0

# SUPABASE関連の追加依存関係
supabase>=2.6.0
google-generativeai>=0.7.0
yt-dlp>=2024.8.6
plotly>=5.17.0
pandas>=2.1.0
altair>=5.0.0
